APLA_Database

APLA Database is the original APLA Dataset

APLA_Database

Format

A data frame with 551 rows and 262 variables:

Q1

Country Coded

Q2

Year Coded

Q5#1_1

Is the country party to the UN Convention relating to Refugees? ('Party to' means ratified) [Yes(1),No(0)]

Q5#2_1

Is the country party to the UN Convention relating to Refugees? ('Party to' means ratified) [Restrictive(0), Liberal(1)]

Q5#3_1_1

Source of Question Q5

Q5#4_1_1

Comments on Question Q5

Q7#1_1

Has the country made any reservations? ('Reservation' are conditions on ratification of the instrument a state can have placed) [Yes(1),No(0)]

Q7#2_1

Has the country made any reservations? ('Reservation' are conditions on ratification of the instrument a state can have placed) [Restrictive(0), Liberal(1)]

Q7#3_1_1

Source of Question Q7

Q7#4_1_1

Comments on Question Q7

Q9#1_1

Is the country party to the UN Convention Against Torture? ('Party to' means ratified) [Yes(1),No(0)]

Q9#2_1

Is the country party to the UN Convention Against Torture? ('Party to' means ratified) [Restrictive(0), Liberal(1)]

Q9#3_1_1

Source of Question Q9

Q9#4_1_1

Comments on Question Q9

Q11#1_1

Is the country party to the Optional Protocol to the Convention Against Torture? ('Party to' means ratified) [Yes(1),No(0)]

Q11#2_1

Is the country party to the Optional Protocol to the Convention Against Torture? ('Party to' means ratified) [Restrictive(0), Liberal(1)]

Q11#3_1_1

Source of Question Q11

Q11#4_1_1

Comments on Question Q11

Q13#1_1

Is the country party to the International Covenant on Civil and Political Rights? ('Party to' means ratified) [Yes(1),No(0)]

Q13#2_1

Is the country party to the International Covenant on Civil and Political Rights? ('Party to' means ratified) [Restrictive(0), Liberal(1)]

Q13#3_1_1

Source of Question Q13

Q13#4_1_1

Comments on Question Q13

Q15#1_1

Is the country party to the First Optional Protocol to the International Covenant on Civil and Political Rights? ('Party to' means ratified) [Yes(1),No(0)]

Q15#2_1

Is the country party to the First Optional Protocol to the International Covenant on Civil and Political Rights? ('Party to' means ratified) [Restrictive(0), Liberal(1)]

Q15#3_1_1

Source of Question Q15

Q15#4_1_1

Comments on Question Q15

Q17#1_1

Is the country party to the Convention on the Rights of the Child? ('Party to' means ratified) [Yes(1),No(0)]

Q17#2_1

Is the country party to the Convention on the Rights of the Child? ('Party to' means ratified) [Restrictive(0), Liberal(1)]

Q17#3_1_1

Source of Question Q17

Q17#4_1_1

Comments on Question Q17

Q19#1_1

Is the country party to the Optional Protcol to the Convention on the Rights of the Child? ('Party to' means ratified. This question relates to the third optional protocol: Optional Protocol to the Convention on the Rights of the Child on a communications Procedures) [Yes(1),No(0)]

Q19#2_1

Is the country party to the Optional Protcol to the Convention on the Rights of the Child? ('Party to' means ratified. This question relates to the third optional protocol: Optional Protocol to the Convention on the Rights of the Child on a communications Procedures) [Restrictive(0), Liberal(1)]

Q19#3_1_1

Source of Question Q19

Q19#4_1_1

Comments on Question Q19

Q200#1_1

Does this track afford protection to persons who may arbitrarily be deprived of their life if returned to their country of origin? [Yes(1),No(0)]

Q200#2_1

Does this track afford protection to persons who may arbitrarily be deprived of their life if returned to their country of origin? [Restrictive(0), Liberal(1)]

Q200#3_1_1

Source of Question Q200

Q200#4_1_1

Comments on Question Q200

Q202#1_1

Does this track afford protection to persons who have been displaced as the result of an ongoing armed conflict within a particular country? [Yes(1),No(0)]

Q202#2_1

Does this track afford protection to persons who have been displaced as the result of an ongoing armed conflict within a particular country? [Restrictive(0), Liberal(1)]

Q202#3_1_1

Source of Question Q202

Q202#4_1_1

Comments on Question Q202

Q204#1_1

Does this track afford protection to persons who have been displaced as the result of a natural/environmental disaster? [Yes(1),No(0)]

Q204#2_1

Does this track afford protection to persons who have been displaced as the result of a natural/environmental disaster? [Restrictive(0), Liberal(1)]

Q204#3_1_1

Source of Question Q204

Q204#4_1_1

Comments on Question Q204

Q278#1_1

Is there any alternative status (subsidiary protection) for persons seeking protection, other than refugee status? [Yes(1),No(0)]

Q278#2_1

Is there any alternative status (subsidiary protection) for persons seeking protection, other than refugee status? [Restrictive(0), Liberal(1)]

Q278#3_1_1

Source of Question Q278

Q278#4_1_1

Comments on Question Q278

Q280#1_1

Does the country provide protection to persons who may be subjected to torture if returned to their country of origin? [Yes(1),No(0)]

Q280#2_1

Does the country provide protection to persons who may be subjected to torture if returned to their country of origin? [Restrictive(0), Liberal(1)]

Q280#3_1_1

Source of Question Q280

Q280#4_1_1

Comments on Question Q280

Q133#1_1

Are asylum seekers detained in some circumstances while their claims are being processed? [Yes(1),No(0)]

Q133#2_1

Are asylum seekers detained in some circumstances while their claims are being processed? [Restrictive(0), Liberal(1)]

Q133#3_1_1

Source of Question Q133

Q133#4_1_1

Comments on Question Q133

Q274#1_1

Are asylum seekers informed of their rights during this screening process? [Yes(1),No(0)]

Q274#2_1

Are asylum seekers informed of their rights during this screening process? [Restrictive(0), Liberal(1)]

Q274#3_1_1

Source of Question Q274

Q274#4_1_1

Comments on Question Q274

Q188#1_1

Are asylum claims that are deemed to be weak or unfounded subject to fast track processing? [Yes(1),No(0)]

Q188#2_1

Are asylum claims that are deemed to be weak or unfounded subject to fast track processing? [Restrictive(0), Liberal(1)]

Q188#3_1_1

Source of Question Q188

Q188#4_1_1

Comments on Question Q188

Q170#1_1

Are asylum seekers advised about their rights in a language they can understand? [i.e. do they have a right to an interpreter?] [Yes(1),No(0)]

Q170#2_1

Are asylum seekers advised about their rights in a language they can understand? [i.e. do they have a right to an interpreter?] [Restrictive(0), Liberal(1)]

Q170#3_1_1

Source of Question Q170

Q170#4_1_1

Comments of Question Q170

Q172#1_1

Are asylum seekers required to submit a written application? [Yes(1),No(0)]

Q172#2_1

Are asylum seekers required to submit a written application? [Restrictive(0), Liberal(1)]

Q172#3_1_1

Source of Question Q172

Q172#4_1_1

Comments on Question Q172

Q174#1_1

Do applicants have a right to an interview/oral hearing? [Yes(1),No(0)]

Q174#2_1

Do applicants have a right to an interview/oral hearing? [Restrictive(0), Liberal(1)]

Q174#3_1_1

Source of Question Q174

Q174#4_1_1

Comments on Question Q174

Q180#1_1

Do applicants have a right to legal assistance or representation? [Yes(1),No(0)]

Q180#2_1

Do applicants have a right to legal assistance or representation? [Restrictive(0), Liberal(1)]

Q180#3_1_1

Source of Question Q180

Q180#4_1_1

Comments on Question Q180

Q184#1_1

Is there a time limit within which asylum seekers must file an application after entering the country? [Yes(1),No(0)]

Q184#2_1

Is there a time limit within which asylum seekers must file an application after entering the country? [Restrictive(0), Liberal(1)]

Q184#3_1_1

Source of Question Q184

Q184#4_1_1

Comments on Question Q184

Q101#1_1

Can refugee protection be denied and/or restricted because an applicant has been in contact with authorities or has remained for a certain period (but not settled) in another country in which they could have sought protection? [Yes(1),No(0)]

Q101#2_1

Can refugee protection be denied and/or restricted because an applicant has been in contact with authorities or has remained for a certain period (but not settled) in another country in which they could have sought protection? [Restrictive(0), Liberal(1)]

Q101#3_1_1

Source of Question Q101

Q101#4_1_1

Comments on Question Q101

Q103#1_1

Can refugee protection be denied to applicants who could avoid persecution by relocating to a different part of their state of origin? [Yes(1),No(0)]

Q103#2_1

Can refugee protection be denied to applicants who could avoid persecution by relocating to a different part of their state of origin? [Restrictive(0), Liberal(1)]

Q103#3_1_1

Source of Question Q103

Q103#4_1_1

Comments on Question Q103

Q107#1_1

Can refugee protection be denied to applicants who have committed crimes against peace, war crimes or crimes against humanity before entry? [Yes(1),No(0)]

Q107#2_1

Can refugee protection be denied to applicants who have committed crimes against peace, war crimes or crimes against humanity before entry? [Restrictive(0), Liberal(1)]

Q107#3_1_1

Source of Question Q107

Q107#4_1_1

Comments on Question Q107

Q113#1_1

Can refugee protection be denied to applicants who represent a danger to national security? [Yes(1),No(0)]

Q113#2_1

Can refugee protection be denied to applicants who represent a danger to national security? [Restrictive(0), Liberal(1)]

Q113#3_1_1

Source of Question Q113

Q113#4_1_1

Comments on Question Q113

Q109#1_1

Can refugee protection be denied to applicants who have committed serious non-political crimes outside the country of refuge (before entry)? [Yes(1),No(0)]

Q109#2_1

Can refugee protection be denied to applicants who have committed serious non-political crimes outside the country of refuge (before entry)? [Restrictive(0), Liberal(1)]

Q109#3_1_1

Source of Question Q109

Q109#4_1_1

Comments on Question Q109

Q253#1_1

Can refugee protection be denied to applicants who have committed serious non-political crimes within the country of refuge? [Yes(1),No(0)]

Q253#2_1

Can refugee protection be denied to applicants who have committed serious non-political crimes within the country of refuge? [Restrictive(0), Liberal(1)]

Q253#3_1_1

Source of Question Q253

Q253#4_1_1

Comments on Question Q253

Q208#1_1

Are asylum seekers whose protection claims are rejected given a time limit to leave the country? [Yes(1),No(0)]

Q208#2_1

Are asylum seekers whose protection claims are rejected given a time limit to leave the country? [Restrictive(0), Liberal(1)]

Q208#3_1_1

Source of Question Q208

Q208#4_1_1

Comments on Question Q208

Q145#1_1

Can child asylum seekers be detained? [Yes(1),No(0)]

Q145#2_1

Can child asylum seekers be detained? [Restrictive(0), Liberal(1)]

Q145#3_1_1

Source of Question Q145

Q145#4_1_1

Comments on Question Q145

Q190#1_1

Are special procedures used or accommodations made for unaccompanied or separated children? [Yes(1),No(0)]

Q190#2_1

Are special procedures used or accommodations made for unaccompanied or separated children? [Restrictive(0), Liberal(1)]

Q190#3_1_1

Source of Question Q190

Q190#4_1_1

Comments on Question Q190

Q194#1_1

Is a guardian appointed to assist the unaccompanied or separated child applicants? [Yes(1),No(0)]

Q194#2_1

Is a guardian appointed to assist the unaccompanied or separated child applicants? [Restrictive(0), Liberal(1)]

Q194#3_1_1

Source of Question Q194

Q194#4_1_1

Comments on Question Q194

Q320#1_1

Is the entrant granted protection with the option of applying for permanent status after a certain period of time? [Yes(1),No(0)]

Q320#2_1

Is the entrant granted protection with the option of applying for permanent status after a certain period of time? [Restrictive(0), Liberal(1)]

Q320#3_1_1

Source of Question Q320

Q320#4_1_1

Comments on Question Q320

Q54#1_1

Does the permit allow the entrant to work? [Yes(1),No(0)]

Q54#2_1

Does the permit allow the entrant to work? [Restrictive(0), Liberal(1)]

Q54#3_1_1

Source of Question Q54

Q54#4_1_1

Comments on Question Q54

Q58#1_1

Can applications be made at the border/ports of entry? [Yes(1),No(0)]

Q58#2_1

Can applications be made at the border/ports of entry? [Restrictive(0), Liberal(1)]

Q58#3_1_1

Source of Question Q58

Q58#4_1_1

Comments on Question Q58

LA1#1_1

Does the law explicitly recognize the duties imposed on the State by the 1951 Geneva convention and other related ratified international treaties? [Yes(1),No(0)]

LA1#2_1

Does the law explicitly recognize the duties imposed on the State by the 1951 Geneva convention and other related ratified international treaties? [Restrictive(0), Liberal(1)]

LA1#3_1_1

Source of Question LA1

LA1#4_1_1

Comments on Question LA1

LA3#1_1

Is the right to asylum included into the Constitution? [Yes(1),No(0)]

LA3#2_1

Is the right to asylum included into the Constitution? [Restrictive(0), Liberal(1)]

LA3#3_1_1

Source of Question LA3

LA3#4_1_1

Comments on Question LA3

LA5#1_1

Does the law explicitly include the possibility to resettle refugees from third countries? [Yes(1),No(0)]

LA5#2_1

Does the law explicitly include the possibility to resettle refugees from third countries? [Restrictive(0), Liberal(1)]

LA5#3_1_1

Source of Question LA5

LA5#4_1_1

Comments on Question LA5

LA7#1_1

Has the country incorporated into its national law the principles of the 1984 Cartagena Declaration on the scope of the refugee definition, or a variation thereof? [Such declaration defines refugees as: persons who have fled their country because their lives, safety or freedom have been threatened by generalized violence, foreign aggression, internal conflicts, massive violation of human rights or other circumstances which have seriously disturbed public order].[Yes(1),No(0)]

LA7#2_1

Has the country incorporated into its national law the principles of the 1984 Cartagena Declaration on the scope of the refugee definition, or a variation thereof? [Such declaration defines refugees as: persons who have fled their country because their lives, safety or freedom have been threatened by generalized violence, foreign aggression, internal conflicts, massive violation of human rights or other circumstances which have seriously disturbed public order]. [Restrictive(0), Liberal(1)]

LA7#3_1_1

Source of Question LA7

LA7#4_1_1

Comments on Question LA7

LA9#1_1

Does the law recognize the right to asylum for people already recognized as refugees in a third country, but whose life, freedom and dignity the host country cannot guarantee? [Yes(1),No(0)]

LA9#2_1

Does the law recognize the right to asylum for people already recognized as refugees in a third country, but whose life, freedom and dignity the host country cannot guarantee? [Restrictive(0), Liberal(1)]

LA9#3_1_1

Source of Question LA9

LA9#4_1_1

Comments on Question LA9

LA11#1_1

Are there any special provisions for cases of mass influx of asylum seekers? [Yes(1),No(0)]

LA11#2_1

Are there any special provisions for cases of mass influx of asylum seekers? [Restrictive(0), Liberal(1)]

LA11#3_1_1

Source of Question LA11

LA11#4_1_1

Comments on Question LA11

LA13#1_1

Does the law guarantee that asylum seekers will not be penalized for entering the country illegally? [Yes(1),No(0)]

LA13#2_1

Does the law guarantee that asylum seekers will not be penalized for entering the country illegally? [Restrictive(0), Liberal(1)]

LA13#3_1_1

Source of Question LA13

LA13#4_1_1

Comments on Question LA13

LA15#1_1

Does the law recognize the declarative character of the refugee condition? (This implies that asylum seekers have the same rights as refugees, as they are refugees waiting to be recognized as such by the host state) [Yes(1),No(0)]

LA15#2_1

Does the law recognize the declarative character of the refugee condition? (This implies that asylum seekers have the same rights as refugees, as they are refugees waiting to be recognized as such by the host state) [Restrictive(0), Liberal(1)]

LA15#3_1_1

Source of Question LA15

LA15#4_1_1

Comments on Question LA15

LA17#1_1

Does the law sanction authorities who fail to pass on to the relevant institution a request for asylum? [That is, does the law mention the duty of public officials to remit a submission for refugee status to the competent authorities?] [Yes(1),No(0)]

LA17#2_1

Does the law sanction authorities who fail to pass on to the relevant institution a request for asylum? [That is, does the law mention the duty of public officials to remit a submission for refugee status to the competent authorities?] [Restrictive(0), Liberal(1)]

LA17#3_1_1

Source of Question LA17

LA17#4_1_1

Comments on Question LA17

LA19#1_1

Can the asylum seeker remain in the country where the application has been made until the last instance of her/his situation has been defined? [Yes(1),No(0)]

LA19#2_1

Can the asylum seeker remain in the country where the application has been made until the last instance of her/his situation has been defined? [Restrictive(0), Liberal(1)]

LA19#3_1_1

Source of Question LA19

LA19#4_1_1

Comments on Question LA19

LA23#1_1

Does the law recognize the unity of the family as a fundamental right of the refugees? [Yes(1),No(0)]

LA23#2_1

Does the law recognize the unity of the family as a fundamental right of the refugees? [Restrictive(0), Liberal(1)]

LA23#3_1_1

Source of Question LA23

LA23#4_1_1

Comments on Question LA23

LA27#1_1

Does the law ease the recognition of academic and professional qualifications earned in the country of origin? [Yes(1),No(0)]

LA27#2_1

Does the law ease the recognition of academic and professional qualifications earned in the country of origin? [Restrictive(0), Liberal(1)]

LA27#3_1_1

Source of Question LA27

LA27#4_1_1

Comments on Question LA27

LA29#1_1

Does the identification document provided to the recognized refugee mention her/his status as a refugee? [Yes(1),No(0)]

LA29#2_1

Does the identification document provided to the recognized refugee mention her/his status as a refugee? [Restrictive(0), Liberal(1)]

LA29#3_1_1

Source of Question LA29

LA29#4_1_1

Comments on Question LA29

LA31#1_1

Does the law guarantee the right to access the asylum process? [Yes(1),No(0)]

LA31#2_1

Does the law guarantee the right to access the asylum process? [Restrictive(0), Liberal(1)]

LA31#3_1_1

Source of Question LA31

LA31#4_1_1

Comments on Question LA31

LA33#1_1

Can the application for refugee status be submitted through the UNHCR? [Yes(1),No(0)]

LA33#2_1

Can the application for refugee status be submitted through the UNHCR? [Restrictive(0), Liberal(1)]

LA33#3_1_1

Source of Question LA33

LA33#4_1_1

Comments on Question LA33

LA35#1_1

Can the application for refugee status be submitted through a legal representative? [Yes(1),No(0)]

LA35#2_1

Can the application for refugee status be submitted through a legal representative? [Restrictive(0), Liberal(1)]

LA35#3_1_1

Source of Question LA35

LA35#4_1_1

Comments on Question LA35

LA37#1_1

Does the law recognize confidentiality as a fundamental part of the refugee application process? [Yes(1),No(0)]

LA37#2_1

Does the law recognize confidentiality as a fundamental part of the refugee application process? [Restrictive(0), Liberal(1)]

LA37#3_1_1

Source of Question LA37

LA37#4_1_1

Comments on Question LA37

LA39#1_1

Does the law guarantee that the lack of documentation will not impede the applicant from submitting her/his request for refugee status? [Yes(1),No(0)]

LA39#2_1

Does the law guarantee that the lack of documentation will not impede the applicant from submitting her/his request for refugee status? [Restrictive(0), Liberal(1)]

LA39#3_1_1

Source of Question LA39

LA39#4_1_1

Comments on Question LA39

LA41#1_1

Does the law include a right to appeal and to a fair trial in case of a first negative decision? [Yes(1),No(0)]

LA41#2_1

Does the law include a right to appeal and to a fair trial in case of a first negative decision? [Restrictive(0), Liberal(1)]

LA41#3_1_1

Source of Question LA41

LA41#4_1_1

Comments on Question LA41

LA43#1_1

Is the institution in charge of reviewing appeals independent from the one of the first instance? [Yes(1),No(0)]

LA43#2_1

Is the institution in charge of reviewing appeals independent from the one of the first instance? [Restrictive(0), Liberal(1)]

LA43#3_1_1

Source of Question LA43

LA43#4_1_1

Comments on Question LA43

LA45#1_1

Does the law guarantee the gratuity of the whole refugee application process? [Yes(1),No(0)]

LA45#2_1

Does the law guarantee the gratuity of the whole refugee application process? [Restrictive(0), Liberal(1)]

LA45#3_1_1

Source of Question LA45

LA45#4_1_1

Comments on Question LA45

LA47#1_1

Does the law guarantee free legal assistance to the asylum seeker? [Yes(1),No(0)]

LA47#2_1

Does the law guarantee free legal assistance to the asylum seeker? [Restrictive(0), Liberal(1)]

LA47#3_1_1

Source of Question LA47

LA47#4_1_1

Comments on Question LA47

LA49#1_1

Does the law state that the interview process should take into account the social and cultural background of the applicant? [Yes(1),No(0)]

LA49#2_1

Does the law state that the interview process should take into account the social and cultural background of the applicant? [Restrictive(0), Liberal(1)]

LA49#3_1_1

Source of Question LA49

LA49#4_1_1

Comments on Question LA49

LA51#1_1

Does the law prohibit national authorities from contacting the country of nationality, or origin, of the applicant, unless when expressly requested by the latter? [Yes(1),No(0)]

LA51#2_1

Does the law prohibit national authorities from contacting the country of nationality, or origin, of the applicant, unless when expressly requested by the latter? [Restrictive(0), Liberal(1)]

LA51#3_1_1

Source of Question LA51

LA51#4_1_1

Comments on Question LA51

LA53#1_1

Does the committee in charge of reviewing and granting refugee status include a member of the UNHCR? [Yes(1),No(0)]

LA53#2_1

Does the committee in charge of reviewing and granting refugee status include a member of the UNHCR? [Restrictive(0), Liberal(1)]

LA53#3_1_1

Source of Question LA53

LA53#4_1_1

Comments on Question LA53

LA55#1_1

Does the committee in charge of reviewing and granting refugee status include a member of the civil society or a representative of the Ombudsman’s Office? [Yes(1),No(0)]

LA55#2_1

Does the committee in charge of reviewing and granting refugee status include a member of the civil society or a representative of the Ombudsman’s Office? [Restrictive(0), Liberal(1)]

LA55#3_1_1

Source of Question LA55

LA55#4_1_1

Comments on Question LA55

LA57#1_1

In case of a negative decision in the last instance of appeal (if applicable), is UNHCR informed of it before any action is undertaken to remove the applicant? [Yes(1),No(0)]

LA57#2_1

In case of a negative decision in the last instance of appeal (if applicable), is UNHCR informed of it before any action is undertaken to remove the applicant? [Restrictive(0), Liberal(1)]

LA57#3_1_1

Source of Question LA57

LA57#4_1_1

Comments on Question LA57

LA59#1_1

In case of rejection in the first instance of the application, is there a reasonable time limit to submit an appeal request? (Reasonable is understood as more than 2 weeks time) [Yes(1),No(0)]

LA59#2_1

In case of rejection in the first instance of the application, is there a reasonable time limit to submit an appeal request? (Reasonable is understood as more than 2 weeks time) [Restrictive(0), Liberal(1)]

LA59#3_1_1

Source of Question LA59

LA59#4_1_1

Comments on Question LA59

LA61#1_1

Are there any special measures in place to guarantee women's access to the asylum procedure? [Yes(1),No(0)]

LA61#2_1

Are there any special measures in place to guarantee women's access to the asylum procedure? [Restrictive(0), Liberal(1)]

LA61#3_1_1

Source of Question LA61

LA61#4_1_1

Comments on Question LA61

LA63#1_1

Are there special provisions in place for vulnerable asylum seekers, or asylum seekers with special needs? [Yes(1),No(0)]

LA63#2_1

Are there special provisions in place for vulnerable asylum seekers, or asylum seekers with special needs? [Restrictive(0), Liberal(1)]

LA63#3_1_1

Source of Question LA63

LA63#4_1_1

Comments on Question LA63

LA65#1_1

Can an individual be recognized as a refugee for persecution based on gender? [Yes(1),No(0)]

LA65#2_1

Can an individual be recognized as a refugee for persecution based on gender? [Restrictive(0), Liberal(1)]

LA65#3_1_1

Source of Question LA65

LA65#4_1_1

Comments on Question LA65

Source

https://github.com/HammoudG/APLA_Dataset