Data_APLA

Data_APLA contains data that includes APLA's regulatory complexity and liberalisation scores, as well as a number of relevant variables that are used in further research.

Data_APLA

Format

A data frame with 551 rows and 33 variables:

Country

character COLUMN_DESCRIPTION

Year

double COLUMN_DESCRIPTION

Art

double COLUMN_DESCRIPTION

Total

double COLUMN_DESCRIPTION

Regulatory_Complexity

double COLUMN_DESCRIPTION

Polity2

double COLUMN_DESCRIPTION

IntMigStock

double COLUMN_DESCRIPTION

Left1_Other0

double COLUMN_DESCRIPTION

Liberalisation

double COLUMN_DESCRIPTION

MigSpain

double COLUMN_DESCRIPTION

MigUS

double COLUMN_DESCRIPTION

MigSpainUS

double COLUMN_DESCRIPTION

TotalPopinY

double COLUMN_DESCRIPTION

MigSpainUSPerc

double COLUMN_DESCRIPTION

GrowthGDPperCap

double COLUMN_DESCRIPTION

Trade_Perc_GDP

double COLUMN_DESCRIPTION

RefugeeAndLikeSit

double COLUMN_DESCRIPTION

TotalPop

double COLUMN_DESCRIPTION

RefAsPerc

double COLUMN_DESCRIPTION

GDPperCapPPP

double COLUMN_DESCRIPTION

Lib100

double COLUMN_DESCRIPTION

x

double COLUMN_DESCRIPTION

y

double COLUMN_DESCRIPTION

CountryID

character COLUMN_DESCRIPTION

South_America

double COLUMN_DESCRIPTION

MigSpainUSLog

double COLUMN_DESCRIPTION

lag1

double COLUMN_DESCRIPTION

lag2

double COLUMN_DESCRIPTION

_est_est1

double COLUMN_DESCRIPTION

_est_est2

double COLUMN_DESCRIPTION

_est_est3

double COLUMN_DESCRIPTION

_est_est4

double COLUMN_DESCRIPTION

VDEM_Polyarchy

double COLUMN_DESCRIPTION

Source

http://somewhere.important.com/